+++ shibboleth.R
@@ -23,40 +23,32 @@ login.shibboleth <- function(){ # Login to ESSE3 via Shibboleth
     e3logurl <- paste0(G$esse3url, "/auth/Logon.do?cod_lingua=eng") # will redir to SSO site
     h <- new_handle()
   
-    ## Get redirected SSO page with sign-in form (usually a different site)
+    ## Request login page. That will redirected to the SSO page (usually on a different site)
     message('.', appendLF = FALSE)
-    SSOresp <- curl_fetch_memory(e3logurl, h)
-    form <- xml_find_all(read_html(rawToChar(SSOresp$content)), "//form")
-    SSOsubmitLink <- xml_attr(form, "action") # the link is relative! 
-    if(!length(SSOsubmitLink)) stop("\nUnable to find the address of login form at ", e3logurl)
-  
-    ## Extract SSO site url from location headers 
-    multihds <- parse_headers(rawToChar(SSOresp$headers), multiple = TRUE)
-    locs <- lapply(multihds, function(hds) grep("^Location: +http", hds, value = TRUE))
-    loc <- sub("^Location: +", "", tail(unlist(locs), 1))
-    SSOendpoint <- regmatches(loc, regexpr("https://.+?/", loc))    
-    SSOpostUrl <- paste0(SSOendpoint, SSOsubmitLink)
+    resp <- curl_fetch_memory(e3logurl, h)
 
-    ## Post Credentials
+    ## Find 'Continue' button and push it ('cause we have no scripts)  
     message('.', appendLF = FALSE)
+    contBtn <- xml_find_all(read_html(rawToChar(resp$content)), "//form//noscript//input[@value='Continue']")
+    form <- xml_find_all(contBtn, "./ancestor::form")
+    resp <- .replayForm(form, resp, h)
+
+    ## Detect credentials' form and post them 
+    message('.', appendLF = FALSE)
+    username.fld <- xml_find_all(read_html(rawToChar(resp$content)), "//form//input[@id='username']") # or @name='j_username'
+    form <- xml_find_all(username.fld, "./ancestor::form")
+    submitLink <- xml_attr(form, "action")
+    url <- paste0(.urlHost(resp$url), submitLink)
     handle_setopt(h,  postfields = creds.post, followlocation = FALSE)
-    SSOresp <- curl_fetch_memory(SSOpostUrl, h)
-    if(!length(SSOresp$content)) stop("\nCredentials might be wrong. Run setCreds()")
+    resp <- curl_fetch_memory(url, h)
+    if(!length(resp$content)) stop("\nCredentials might be wrong. Run setCreds()")###########3
 
-    ## Get the "continue" form data, and simulate click to continue ('cause we have no scripts)
-    form <- xml_find_all(read_html(rawToChar(SSOresp$content)), "//form") 
-    if(!length(form)) stop("Unable to post credentials to\n", SSOpostUrl)
-    relaystate <- xml_attr(xml_find_all(form, ".//input[@name='RelayState']"), "value") 
-    samlresp <- xml_attr(xml_find_all(form, ".//input[@name='SAMLResponse']"), "value") 
-    samlresp <- curl_escape(samlresp)
-    postData <- charToRaw(paste0("RelayState=", relaystate, "&SAMLResponse=", samlresp, "&_eventId_proceed="))
-
-    ## Submit "continue"
+    ## Find 'Continue' button and push it ('cause we have no scripts)  
     message('.', appendLF = FALSE)
-    handle_setopt(h, postfields = postData, followlocation = FALSE)
-    logurl <- xml_attr(form, "action")
-    logresp <- curl_fetch_memory(logurl, h)
-
+    contBtn <- xml_find_all(read_html(rawToChar(resp$content)), "//form//noscript//input[@value='Continue']")
+    form <- xml_find_all(contBtn, "./ancestor::form")
+    resp <- .replayForm(form, resp, h)
+        
     ## Test Shibboleth and Esse3
     if(!length(grep("_shibsession_", handle_cookies(h)$name))){
         stop("\nI did not get expected cookies at ", logurl)
@@ -150,4 +142,4 @@ login.moodle.shibboleth <- function(){ # Login to Moodle via Shibboleth
 
 
 ### Please do not touch the strng below
-### nC2L35TPRsbXlqUw9Ku$GV@wrzz34$Kk*eSw#Q8LlbafywtE*bT3QGn%
+### 237f151825685a480373724d0f2d4b583078675167553d3823114f302e